+++ tests/lightning_tensor/test_lightning_tensor.py
@@ -71,7 +71,5 @@ def test_invalid_keyword_arg():
 
 def test_invalid_shots():
     """Test that an error is raised if finite number of shots are requestd."""
-    with pytest.raises(
-        ValueError, match="LightningTensor does not support the `shots` parameter."
-    ):
+    with pytest.raises(ValueError, match="LightningTensor does not support the `shots` parameter."):
         LightningTensor(shots=5)